+++ tests/test_core.py
@@ -144,6 +144,7 @@ class TestCore(unittest.TestCase):
                 "░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░  0.0%",
                 end="\r",
             ),
+            mock.call("\033[K", end=""),
             mock.call("/tmp/entomb_testing/subdirectory/immutable.txt"),
             mock.call("\033[K", end=""),
             mock.call(
@@ -212,18 +213,21 @@ class TestCore(unittest.TestCase):
                 "██████████░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░  25.0%",
                 end="\r",
             ),
+            mock.call("\033[K", end=""),
             mock.call("/tmp/entomb_testing/mutable.txt"),
             mock.call("\033[K", end=""),
             mock.call(
                 "███████████████░░░░░░░░░░░░░░░░░░░░░░░░░  37.5%",
                 end="\r",
             ),
+            mock.call("\033[K", end=""),
             mock.call("/tmp/entomb_testing/.git/mutable.txt"),
             mock.call("\033[K", end=""),
             mock.call(
                 "████████████████████░░░░░░░░░░░░░░░░░░░░  50.0%",
                 end="\r",
             ),
+            mock.call("\033[K", end=""),
             mock.call("/tmp/entomb_testing/.git/subdirectory/mutable.txt"),
             mock.call("\033[K", end=""),
             mock.call(
@@ -240,6 +244,7 @@ class TestCore(unittest.TestCase):
                 "███████████████████████████████████░░░░░  87.5%",
                 end="\r",
             ),
+            mock.call("\033[K", end=""),
             mock.call("/tmp/entomb_testing/subdirectory/mutable.txt"),
             mock.call("\033[K", end=""),
             mock.call(